requirements.txt
alembic==1.11.1
apache-airflow==2.6.2
apache-airflow-providers-amazon==5.1.0
boto3==1.17.39
botocore==1.20.11
gunicorn==20.1.0
Jinja2==3.1.2
virtualenv==20.13.3
python-dotenv==0.21.1
pydantic>=1.10.0,<2.0.0
pandas==2.0.2
mysql-connector-python==8.1.0
